Removed command line usage of maven and git
  - added methods to compute default config values
  - made libero stateless
  - separated cli from libero-core and reorganized projects
  - Moved libera under subprojects directory
  - Moved Options to Libero and refactored
  - Added deploy to build tasks
