+++ README.md
@@ -1,2 +1,5 @@
-# libero [![Build Status](http://img.shields.io/travis/connexta/libero/master.svg?style=flat-square)](https://travis-ci.org/connexta/libero)
+# libero 
+[![Build Status](http://img.shields.io/travis/connexta/libero/master.svg?style=flat-square)](https://travis-ci.org/connexta/libero)
+[![GitHub license](https://img.shields.io/github/license/connexta/libero.svg?style=flat-square)](https://github.com/connexta/libero/blob/master/LICENSE)
+
 automating releases of git repositories
