--- libero/build.gradle
@@ -1,14 +1,0 @@
-apply plugin: 'application'
-
-group = 'com.connexta.libero'
-version = '1.0.0-SNAPSHOT'
-
-mainClassName = 'com.connexta.libero.Cli'
-
-dependencies {
-    compile 'org.codehaus.groovy:groovy-all:2.4.4'
-    compile 'org.jyaml:jyaml:1.3'
-    compile 'commons-cli:commons-cli:1.3.1'
-    testCompile 'org.spockframework:spock-core:1.0-groovy-2.4'
-    testCompile 'cglib:cglib-nodep:3.2.2'
-}
