+++ subprojects/libera/build.gradle
@@ -1,0 +1,16 @@
+apply plugin: 'application'
+
+group = 'com.connexta.libera'
+version = '1.0.0-SNAPSHOT'
+
+mainClassName = 'com.connexta.libera.Libera'
+
+dependencies {
+    compile 'org.codehaus.groovy:groovy-all:2.4.4'
+    compile 'org.jyaml:jyaml:1.3'
+    compile 'commons-cli:commons-cli:1.3.1'
+    compile project(':libero-core')
+    testCompile 'org.spockframework:spock-core:1.0-groovy-2.4'
+    testCompile 'cglib:cglib-nodep:3.2.2'
+    testCompile 'org.objenesis:objenesis:2.2'
+}
