+++ subprojects/libero-core/build.gradle
@@ -1,0 +1,20 @@
+apply plugin: 'groovy'
+
+group = 'com.connexta.libero.core'
+
+println uploadArchives.toString()
+
+dependencies {
+    compile 'org.codehaus.groovy:groovy-all:2.4.4'
+    compile 'org.jyaml:jyaml:1.3'
+    compile 'commons-cli:commons-cli:1.3.1'
+    compile 'org.apache.maven.shared:maven-invoker:2.2'
+    compile 'org.ajoberstar:grgit:1.6.0'
+    testCompile 'org.spockframework:spock-core:1.0-groovy-2.4'
+}
+
+test {
+    testLogging {
+        exceptionFormat = 'full'
+    }
+}
