+++ subprojects/libero-core/src/main/groovy/com/connexta/libero/Options.groovy
@@ -1,0 +1,25 @@
+package com.connexta.libero
+
+/**
+ * Class that encapsulates the release execution options.
+ */
+class Options {
+    final boolean gitPush
+    final boolean dryRun
+    final boolean force
+    final boolean quickBuild
+
+    Options(map) {
+        gitPush = map.gitPush
+        dryRun = map.dryRun
+        force = map.force
+        quickBuild = map.quickBuild
+    }
+
+    Options(OptionAccessor options) {
+        this.gitPush = options.p ?: false
+        this.dryRun = options.t ?: false
+        this.force = options.f ?: false
+        this.quickBuild = options.q ?: false
+    }
+}
