+++ subprojects/libero-core/src/test/resources/gitRepo/README.md
@@ -1,0 +1,1 @@
+# Test git repository files
