+++ subprojects/libero-core/src/test/resources/gitRepo/branchFile.txt
@@ -1,0 +1,1 @@
+This file will be added to a test branch
