+++ subprojects/libero-distribution/build.gradle
@@ -1,0 +1,21 @@
+apply plugin: 'groovy'
+apply plugin: 'application'
+apply plugin: 'maven'
+apply plugin: 'build-dashboard'
+
+group = 'com.connexta.libero.distribution'
+version = '1.0.0-SNAPSHOT'
+
+applicationName = 'libero'
+
+mainClassName = 'com.connexta.libero.cli.Cli'
+
+dependencies {
+    compile project(':libero-core')
+}
+
+test {
+    testLogging {
+        exceptionFormat = 'full'
+    }
+}
